==> 01_plotting.Rmd <==
---
title: "Examining the crash data"
output:
  html_document: 
    toc: true
    toc_depth: 2
    toc_float: true
    theme: cosmo
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

##Normalizing the data

One of the concerns we had is that we would basically be looking at towns with high populations - the higher the population, the more pedestrian and cyclist accidents there would be.

We used the population data for each year to compute the rate of cyclist crashes, pedestrican crashes and total crashes per 1,000 population. 

To do this, according to Numbers in the Newsroom (available [here](https://store.ire.org/products/numbers-in-the-newsroom-using-math-and-statistics-in-news-second-edition-e-version) from the IRE), you take the number of crashes and divide by the population. Then multiply that number by 1000.

This way we can compare more than just the raw numbers. And, we can get a bit more of an apples-to-apples comparison between large and small towns.

##Working in R

Start with loading the libraries we need along with the data:

```{r}

# load the libraries 
library(readr)
library(ggplot2)
library(ggthemes)
#-------------------

df <- read_csv("data/all_crashData_0809.csv")


```

Select everything except Chicago as those totals are far higher than our surburban numbers, which will skew our visualizations.

```{r}

df <- df[df$town != "Chicago",]

```


##Charting density

Summary gives us a look at each column, including the mean and median. In case we want to reference that later, we store the results of that summary call in a dataframe.

```{r}

dfsum <- data.frame(unclass(summary(df)), check.names = FALSE, stringsAsFactors = FALSE)
dfsum

```

Here's a density chart on total pedestrian crashes, by year

```{r}

qplot(ped_total, data=df, 
      geom="density",
      fill = factor(year),
      color = factor(year)
      #alpha=I(.5) 
) +
  facet_grid(. ~ factor(year))

```

We also looked at charts for cyclist crashes and total crashes, as well as the rates for these. For the most part, the shape of the data was about the same - many places had a few amounts of each and a few places had many amounts.

##Comparing to population

We used scatter plots to compare to total population. Here's total cycle crashes vs. population, by year.

```{r}

qplot(pop,cycle_total,
      data=df,
      alpha=I(.25)) + 
  stat_smooth(method="lm") +
  facet_grid(. ~ factor(year))

```

As expected, places with higher populations generally had higher numbers of cycle crashes.

Comparing the crash rate per 1000 yielded different results. 

```{r}

qplot(pop,cycle_rate,
      data=df,
      alpha=I(.25)) + 
  stat_smooth(method="lm") +
  facet_grid(. ~ factor(year))

```


That's hard to read. There's a few towns with very small populations where one crash yields a very high rate, and that's misleading.

Let's exclude rates above 2 since so we can see the results better.

```{r}

dfa <- df[df$total_rate < 2,]

qplot(pop,cycle_rate,
      data=dfa,
      alpha=I(.25)) + 
  stat_smooth(method="lm") +
  facet_grid(. ~ factor(year))

```

Here's pedestrian crashes vs. pop.

```{r}

qplot(pop,ped_total,
      data=dfa,
      alpha=I(.25)) + 
  stat_smooth(method="lm") +
  facet_grid(. ~ factor(year))

```

Here's the rate 1000 for pedestrian crashes vs pop.

```{r}

qplot(pop,ped_rate,
      data=dfa,
      alpha=I(.25)) + 
  stat_smooth(method="lm") +
  facet_grid(. ~ factor(year))

```

Here's total crashes vs. pop.

```{r}

qplot(pop,total_crashes,
      data=dfa,
      alpha=I(.25)) + 
  stat_smooth(method="lm") +
  facet_grid(. ~ factor(year))

```

Here's the rate per 1000 of total crashes vs. pop.

```{r}

qplot(pop,total_rate,
      data=dfa,
      alpha=I(.25)) + 
  stat_smooth(method="lm") +
  facet_grid(. ~ factor(year))

```

So what are we seeing here?
Places with higher populations tend to have more crashes.  
But even within that expected result we can see places with populations over 50,000 with higher than expected rate of crashes. 


##Injuries per 1000

Here's number of injuries compared to crashes per 1000


```{r}

qplot(cycle_rate,cycle_inj,
      data=dfa,
      alpha=I(.25)) + 
  stat_smooth(method="lm") +
  facet_grid(. ~ factor(year))

```

```{r}

qplot(ped_rate,ped_inj,
      data=dfa,
      alpha=I(.25)) + 
  stat_smooth(method="lm") +
  facet_grid(. ~ factor(year))

```

```{r}

qplot(total_rate,total_inj,
      data=dfa,
      alpha=I(.25)) + 
  stat_smooth(method="lm") +
  facet_grid(. ~ factor(year))

```

Generally it looks like the higher the crash rate, the more injuries - which might be expected. 
Places with a relatively low number of injuries but with a high rate of injuries are likely areas of low population.
But there looks to be many places with higher than expected injuries vs. their rate.
