==> 02_Map_prep.Rmd <==
---
title: "Prepping the data for mapping"
output:
  html_document: 
    toc: true
    toc_depth: 2
    toc_float: true
    theme: cosmo

---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

In order to view the data geographically, or on a map, it needs to be matched up with geographies in the form of shapefiles. Shapefiles are the electronic files that contain not just the shape of a place, but also any data that's associated with that place.

First we load in the libraries we need and our data.

```{r}

library(readr)
library(ggplot2)
library(rgdal)
library(dplyr)
library(ggmap)

df <- read_csv("data/all_crashData_0809.csv")

```


##Working with shapefiles


```{r}

ill <- readOGR(dsn = "sixCoplaces/smaller2.shp")

```

smaller2.shp contains places - meaning towns and other municipalities - in the six-county area. Shapefiles can be very large and detailed but the shapes in this have been simplified using a mapping program called Qgis. By simplified, we mean the shapes have been reduced in complexity - they have fewer points, which keeps the file size down.

The shapes aren't going to be accurate to the square foot, but they're still representative of the municipality.

Each place in the shapefile is associated with information. That information in the file can be accessed using the "@" sign instead of a "$" for dataframes. And, it's called "data".

```{r}

head(ill@data, n = 10)
summary(ill@data)

```

We can see the shapes using the "plot" command.

```{r}

plot(ill)

```

##Checking the data

We can test how well our data will match up with the data in the shapefile. We'll see if the town names in ill@data match up with the town names in our dataframe.

```{r}

ill$NAME10 %in% df$town

```

That results in a lot of "false" results, where there was no match for a place.

We ended up exporting the data and looking at it in excel. We found that all the towns in our data set match up with the towns in the shapefile. The false results were towns for which we had no data.

##Reworking the data for mapping

The next problem is reworking our data so we can match it up a town. 

In our dataframe, each town has up to four rows - one for each year. We need each town to have only one row to match up with the one geography.

First we grabbed the data by year

```{r}

df2012 <- df[df$year == 2012,]
df2013 <- df[df$year == 2013,]
df2014 <- df[df$year == 2014,]
df2015 <- df[df$year == 2015,]

```

Then we renamed the columns so their names aren't duplicated. We also deleted or nulled some unneeded columns.

```{r}

df2012 <- rename(df2012, 
       cycle_total_2012 = cycle_total,
       cycle_fatal_2012 = cycle_fatal,
       cycle_inj_2012 = cycle_inj,
       ped_total_2012 = ped_total,
       ped_fatal_2012 = ped_fatal,
       ped_inj_2012 = ped_inj,
       total_crashes_2012 = total_crashes,
       total_inj_2012 = total_inj,
       total_fatal_2012 = total_fatal,
       pop_2012 = pop,
       cycle_rate_2012 = cycle_rate,
       ped_rate_2012 = ped_rate,
       total_rate_2012 = total_rate
       )

df2012$town2 = NULL
df2012$year = NULL

df2013 <- rename(df2013, 
                 cycle_total_2013 = cycle_total,
                 cycle_fatal_2013 = cycle_fatal,
                 cycle_inj_2013 = cycle_inj,
                 ped_total_2013 = ped_total,
                 ped_fatal_2013 = ped_fatal,
                 ped_inj_2013 = ped_inj,
                 total_crashes_2013 = total_crashes,
                 total_inj_2013 = total_inj,
                 total_fatal_2013 = total_fatal,
                 pop_2013 = pop,
                 cycle_rate_2013 = cycle_rate,
                 ped_rate_2013 = ped_rate,
                 total_rate_2013 = total_rate
)
df2013$town2 = NULL
df2013$year = NULL

df2014 <- rename(df2014, 
                 cycle_total_2014 = cycle_total,
                 cycle_fatal_2014 = cycle_fatal,
                 cycle_inj_2014 = cycle_inj,
                 ped_total_2014 = ped_total,
                 ped_fatal_2014 = ped_fatal,
                 ped_inj_2014 = ped_inj,
                 total_crashes_2014 = total_crashes,
                 total_inj_2014 = total_inj,
                 total_fatal_2014 = total_fatal,
                 pop_2014 = pop,
                 cycle_rate_2014 = cycle_rate,
                 ped_rate_2014 = ped_rate,
                 total_rate_2014 = total_rate
)
df2014$town2 = NULL
df2014$year = NULL

df2015 <- rename(df2015, 
                 cycle_total_2015 = cycle_total,
                 cycle_fatal_2015 = cycle_fatal,
                 cycle_inj_2015 = cycle_inj,
                 ped_total_2015 = ped_total,
                 ped_fatal_2015 = ped_fatal,
                 ped_inj_2015 = ped_inj,
                 total_crashes_2015 = total_crashes,
                 total_inj_2015 = total_inj,
                 total_fatal_2015 = total_fatal,
                 pop_2015 = pop,
                 cycle_rate_2015 = cycle_rate,
                 ped_rate_2015 = ped_rate,
                 total_rate_2015 = total_rate
)
df2015$town2 = NULL
df2015$year = NULL

```


With that, we joined our dated columns to the data in the shapefile. We also deleted some unneeded columns from the shapefile, saved the data and then read it back in.


```{r}

ill@data <- left_join(ill@data, df2012, by = c('NAME10' = 'town'))
ill@data <- left_join(ill@data, df2013, by = c('NAME10' = 'town'))
ill@data <- left_join(ill@data, df2014, by = c('NAME10' = 'town'))
ill@data <- left_join(ill@data, df2015, by = c('NAME10' = 'town'))

ill@data$POPESTIMAT <- NULL
ill@data$POPESTIM_1 <- NULL
ill@data$POPESTIM_2 <- NULL
ill@data$change2010 <- NULL
ill@data$change2014 <- NULL
ill@data$change20_1 <- NULL
names(ill)

write_csv(ill@data,"data/mapdata.csv")
cycped <- read.csv("data/mapdata.csv", stringsAsFactors = FALSE)
head(cycped)

```


##Creating the dataframe for mapping

In order to use the shapefile with R's ggplot2 graphical package, we need to process it. 

```{r}

ill_f <- fortify(ill, region="GEOID10")

```


Once we do that, we need to rejoin the data we saved to the new dataframe called ill_f. 

To do that, the GEOID number in the new dataframe needs to be made a number, so that we can join it with the GEOID in our data file, which is already a recognized as a number.

Then we can write the dataframe out to a csv to use for mapping.

```{r}

ill_f$id <- as.numeric(as.character(ill_f$id))

ill_f <- left_join(ill_f, cycped, by = c('id' = 'GEOID10'))

write_csv(ill_f,"data/ill_f.csv")

```

Finally, we import and process a shapefile of the six counties. We'll put this on top of the places layer of any map we do to make it easier to recognize where things are.

```{r}

counties <- readOGR(dsn = "sixcounties/SixCounties.shp")
names(counties@data)
co6 <- fortify(counties, region="geoid10")
write_csv(co6,"data/counties6.csv")

```
